refactor: exclusion of closed tag items in component completion
If coc-html is installed, it will give you closing tag suggestions

coc-laravel uses vscode-html-languageservice to provide additional
tag-related completions.

It seems that as many closing tags as the number of html-tag related
completion feature added in coc-laravel are duplicated and displayed as
items.

Therefore, the html-tag related completion on the coc-laravel side has
filtered out items with closing tags