refactor!: remove Blade directive hover feature
This is a feature ported from coc-blade.

In providing hover, we have included a number of markdown files with the
extension, Since the size of coc-laravel's own package has recently
grown, we will remove the markdown files and other files related to
blade directive hover.

Therefore, the blade directive hover function itself is also removed.