refactor!: stop watch-files under vendor/composer
Currently, files under vendor/composer are monitored and the process of
restarting the project is triggered when a change is detected.

When a package is installed with composer, many files are updated, and
this causes the restart process to run a lot, making the PC very
overloaded.

This is not very good, so we stop monitoring under vendor/composer once.
Consider alternative approaches...