perf: improve the speed of displaying complementary items
Slow display of completion items due to the addition of many completion
functions. This is noticeable when the file size is large in blade
files.

I think this is because there is a process within each handler that
decides whether to provide a completed item based on the current cursor
position, and that is done for each handler.

Added region determination process to the main provideCompletionItems
side. For now, we will leave the processing in each handler as it is,
but we will consider adjustments there as well in the future.